pet duration to minutes

+++ app/src/main/java/com/desireProj/ble_sdk/database/ETLItem.kt
@@ -3,10 +3,10 @@ package com.desireProj.ble_sdk.database
 class ETLItem {
     val pet: String
     val day: String
-    val duration: Long
+    val duration: Float
     val rssi: Int
 
-    constructor(pet: String, day: String, duration: Long, rssi: Int) {
+    constructor(pet: String, day: String, duration: Float, rssi: Int) {
         this.pet = pet
         this.day = day
         this.duration = duration
